+++ packages/varlet-ui/src/card/example/locale/zh-CN.ts
@@ -1,0 +1,10 @@
+export default {
+  basicUsage: '基本使用',
+  title: '这是标题',
+  showSubtitle: '显示副标题',
+  subtitle: '这是副标题',
+  description: '这是描述',
+  showImage: '显示图片',
+  useSlot: '使用插槽',
+  button: '添加按钮',
+}