+++ packages/varlet-ui/src/snackbar/index.ts
@@ -40,21 +40,21 @@ interface UniqSnackbarOptions {
 }
 
 interface Snackbar {
-  (options: SnackbarOptions | string): SnackbarHandel
+  (options: any): SnackbarHandel
 
   install(app: App): void
 
   allowMultiple(bool: boolean): void
 
-  success(options: SnackbarOptions | string): SnackbarHandel
+  success(options: any): SnackbarHandel
 
-  warning(options: SnackbarOptions | string): SnackbarHandel
+  warning(options: any): SnackbarHandel
 
-  info(options: SnackbarOptions | string): SnackbarHandel
+  info(options: any): SnackbarHandel
 
-  error(options: SnackbarOptions | string): SnackbarHandel
+  error(options: any): SnackbarHandel
 
-  loading(options: SnackbarOptions | string): SnackbarHandel
+  loading(options: any): SnackbarHandel
 
   clear(): void
 
@@ -67,6 +67,21 @@ let sid = 0
 let isMount = false
 let unmount: () => void
 let uniqSnackbarOptions: Array<UniqSnackbarOptions> = reactive<UniqSnackbarOptions[]>([])
+
+const defaultOption: Partial<Record<keyof SnackbarOptions, any>> = {
+  position: 'top',
+  duration: 3000,
+  vertical: false,
+  loadingType: 'circle',
+  loadingSize: 'normal',
+  lockScroll: false,
+  teleport: 'body',
+  forbidClick: false,
+  onOpen: () => {},
+  onOpened: () => {},
+  onClose: () => {},
+  onClosed: () => {},
+}
 
 const transitionGroupProps: any = {
   name: 'var-snackbar-fade',
@@ -94,9 +109,11 @@ const TransitionGroupHost = {
       })
 
       const snackbarList = uniqSnackbarOptions.map((option: UniqSnackbarOptions) => {
-        if (option.reactiveSnackOptions.forbidClick) {
-          const transitionGroupEl = document.querySelector('.var-transition-group')
+        const transitionGroupEl = document.querySelector('.var-transition-group')
+        if (option.reactiveSnackOptions.forbidClick || option.reactiveSnackOptions.type === 'loading') {
           ;(transitionGroupEl as HTMLElement).classList.add('var-pointer-auto')
+        } else {
+          ;(transitionGroupEl as HTMLElement).classList.remove('var-pointer-auto')
         }
 
         if (Snackbar.isAllowMultiple) option.reactiveSnackOptions.position = 'top'
@@ -123,9 +140,12 @@ const TransitionGroupHost = {
   },
 }
 
-const Snackbar: Snackbar = <Snackbar>function (options: SnackbarOptions | string): SnackbarHandel {
-  const snackOptions: SnackbarOptions = isPlainObject(options) ? options : { content: options }
-  const reactiveSnackOptions: SnackbarOptions = reactive<SnackbarOptions>(snackOptions)
+const Snackbar: Snackbar = <Snackbar>function (options: any): SnackbarHandel {
+  const snackOptions: SnackbarOptions = isPlainObject(options) ? options : { content: String(options) }
+  const reactiveSnackOptions: SnackbarOptions = reactive<SnackbarOptions>({
+    ...defaultOption,
+    ...snackOptions
+  })
   reactiveSnackOptions.show = true
 
   if (!isMount) {
@@ -158,14 +178,14 @@ const Snackbar: Snackbar = <Snackbar>function (options: SnackbarOptions | string
 }
 
 SNACKBAR_TYPE.forEach((type) => {
-  Snackbar[type] = (options: SnackbarOptions | string): SnackbarHandel => {
-    if (typeof options === 'string') {
+  Snackbar[type] = (options: any): SnackbarHandel => {
+    if (isPlainObject(options)) {
+      options.type = type
+    } else {
       options = {
-        content: options,
+        content: String(options),
         type,
       }
-    } else {
-      options.type = type
     }
     return Snackbar(options)
   }